+++ packages/svelte/formula/src/types/options.ts
@@ -1,3 +1,5 @@
+import { CustomValidationRules } from './validation';
+
 /**
  * Optional settings for Formula
  */
@@ -6,4 +8,8 @@ export interface FormulaOptions {
    * Override locale for sorting
    */
   locale?: string;
+  /**
+   * Custom Validators for fields
+   */
+  validators?: CustomValidationRules;
 }